create_repo/requirements.txt
requests
PyYAML
zstandard